refactor: move stateless utilities to DIRACCommon
Move ClassAd, ParametricJob, and stateless JobDBUtils functions to DIRACCommon
to reduce circular dependencies with diracx. Add backward compatibility
re-exports in DIRAC modules.

Changes:
- DIRACCommon/Utils/ClassAd/ClassAdLight.py: Complete ClassAd implementation
- DIRACCommon/Utils/ParametricJob.py: Parametric job utilities
- DIRACCommon/Utils/JobDBUtils.py: Stateless JDL compression/decompression
- Added comprehensive tests for all moved functionality
- Updated DIRAC modules to re-export from DIRACCommon for compatibility
