+++ dirac-common/src/DIRACCommon/Utils/ClassAd/__init__.py
@@ -1,0 +1,1 @@
+"""ClassAd utilities for DIRACCommon"""
