+++ dirac-common/tests/Utils/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon Utils tests"""
