+++ src/DIRAC/WorkloadManagementSystem/Utilities/test/Test_ParametricJob.py
@@ -1,138 +1,55 @@
-""" This is a test of the parametric job generation tools
-"""
+""" This is a test of the parametric job generation tools"""
 # pylint: disable= missing-docstring
 
 import pytest
 
+# Test imports from DIRAC to verify backward compatibility
 from DIRAC.WorkloadManagementSystem.Utilities.ParametricJob import generateParametricJobs, getParameterVectorLength
 from DIRAC.Core.Utilities.ClassAd.ClassAdLight import ClassAd
 
-TEST_JDL_NO_PARAMETERS = """
-[
-    Executable = "my_executable";
-    Arguments = "%s";
-    JobName = "Test_%n";
-]
-"""
 
-TEST_JDL_SIMPLE = """
-[
-    Executable = "my_executable";
-    Arguments = "%s";
-    JobName = "Test_%n";
-    Parameters = { "a", "b", "c" }
-]
-"""
-
-TEST_JDL_SIMPLE_BUNCH = """
-[
-    Executable = "my_executable";
-    Arguments = "%s";
-    JobName = "Test_%n";
-    Parameters = 3;
-    ParameterStart = 5;
-]
-"""
-
-TEST_JDL_SIMPLE_PROGRESSION = """
-[
-    Executable = "my_executable";
-    Arguments = "%s";
-    JobName = "Test_%n";
-    Parameters = 3;
-    ParameterStart = 1;
-    ParameterStep = 1;
-    ParameterFactor = 2;
-]
-"""
-
-TEST_JDL_MULTI = """
-[
-    Executable = "my_executable";
-    Arguments = "%(A)s %(B)s";
-    JobName = "Test_%n";
-    Parameters = 3;
-    ParameterStart.A = 1;
-    ParameterStep.A = 1;
-    ParameterFactor.A = 2;
-    Parameters.B = { "a","b","c" };
-]
-"""
-
-TEST_JDL_MULTI_BAD = """
-[
-    Executable = "my_executable";
-    Arguments = "%(A)s %(B)s";
-    JobName = "Test_%n";
-    Parameters = 3;
-    ParameterStart.A = 1;
-    ParameterStep.A = 1;
-    ParameterFactor.A = 2;
-    Parameters.B = { "a","b","c","d" };
-]
-"""
-
-
-@pytest.mark.parametrize(
-    "jdl, expectedArguments",
+def test_backward_compatibility_import():
+    """Test that imports from DIRAC still work (backward compatibility)"""
+    # Arrange
+    jdl = """
     [
-        (TEST_JDL_SIMPLE, ["a", "b", "c"]),
-        (TEST_JDL_SIMPLE_BUNCH, ["5", "5", "5"]),
-        (TEST_JDL_SIMPLE_PROGRESSION, ["1", "3", "7"]),
-        (TEST_JDL_MULTI, ["1 a", "3 b", "7 c"]),
-        (TEST_JDL_NO_PARAMETERS, []),
-    ],
-)
-def test_getParameterVectorLength_successful(jdl: str, expectedArguments: list[str]):
-    # Arrange
+        Executable = "my_executable";
+        Arguments = "%s";
+        JobName = "Test_%n";
+        Parameters = { "a", "b", "c" }
+    ]
+    """
+    
+    # Act - Test that we can import and use the functions from DIRAC
     jobDescription = ClassAd(jdl)
-
-    # Act
-    result = getParameterVectorLength(jobDescription)
-
-    # Assert
-    assert result["OK"], result["Message"]
-    if expectedArguments:
-        assert result["Value"] == len(expectedArguments)
-    else:
-        assert result["Value"] == None
+    vector_result = getParameterVectorLength(jobDescription)
+    generate_result = generateParametricJobs(jobDescription)
+    
+    # Assert - Verify functions work correctly
+    assert vector_result["OK"]
+    assert vector_result["Value"] == 3
+    assert generate_result["OK"]
+    assert len(generate_result["Value"]) == 3
 
 
-@pytest.mark.parametrize("jdl", [TEST_JDL_MULTI_BAD])
-def test_getParameterVectorLength_unsuccessful(jdl: str):
-    # Arrange
-    jobDescription = ClassAd(jdl)
-
-    # Act
-    result = getParameterVectorLength(jobDescription)
-
-    # Assert
-    assert not result["OK"], result["Value"]
-
-
-@pytest.mark.parametrize(
-    "jdl, expectedArguments",
-    [
-        (TEST_JDL_SIMPLE, ["a", "b", "c"]),
-        (TEST_JDL_SIMPLE_BUNCH, ["5", "5", "5"]),
-        (TEST_JDL_SIMPLE_PROGRESSION, ["1", "3", "7"]),
-        (TEST_JDL_MULTI, ["1 a", "3 b", "7 c"]),
-    ],
-)
-def test_generateParametricJobs(jdl: str, expectedArguments: list[str]):
-    # Arrange
-    parametricJobDescription = ClassAd(jdl)
-
-    # Act
-    result = generateParametricJobs(parametricJobDescription)
-
-    # Assert
-    assert result["OK"], result["Message"]
-    assert result["Value"]
-    jobDescList = result["Value"]
-    assert len(jobDescList) == len(expectedArguments)
-
-    for i in range(len(jobDescList)):
-        jobDescription = ClassAd(jobDescList[i])
-        assert jobDescription.getAttributeString("JobName") == f"Test_{i}"
-        assert jobDescription.getAttributeString("Arguments") == expectedArguments[i]
+# Import and run the comprehensive tests from DIRACCommon to avoid duplication
+# This ensures the DIRAC re-exports work with the full test suite
+try:
+    from DIRACCommon.tests.Utils.test_ParametricJob import (
+        test_getParameterVectorLength_successful,
+        test_getParameterVectorLength_unsuccessful, 
+        test_generateParametricJobs,
+    )
+    
+    # Re-export the DIRACCommon tests so they run as part of DIRAC test suite
+    # This validates that the backward compatibility imports work correctly
+    __all__ = [
+        'test_backward_compatibility_import',
+        'test_getParameterVectorLength_successful',
+        'test_getParameterVectorLength_unsuccessful',
+        'test_generateParametricJobs',
+    ]
+    
+except ImportError:
+    # If DIRACCommon tests can't be imported, just run the backward compatibility test
+    __all__ = ['test_backward_compatibility_import']
